requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Data Validation
pydantic>=2.4.0
pydantic[email]>=2.4.0
pydantic-settings>=2.0.0

# External Services
resend>=0.6.0

# Configuration
python-dotenv>=1.0.0
